+++ packages/happner-cluster/test/_lib/integration-36/remote-component-2/index.js
@@ -1,0 +1,20 @@
+module.exports = class Component {
+  constructor() {
+    this.version = '3.4.5';
+  }
+  start($happn, callback) {
+    callback();
+  }
+
+  stop($happn, callback) {
+    callback();
+  }
+
+  getVersion($happn, callback) {
+    callback(null, {
+      mesh: $happn.info.mesh.name,
+      version: this.version,
+      component: 'remoteComponent2',
+    });
+  }
+}