+++ RELEASES.md
@@ -4,3 +4,10 @@ happner-2: 12.1.0
 happner-client: 12.1.0
 happner-cluster: 12.1.0
 
+1.2.0 2022-04-12
+-----------------
+happner-2: 12.1.1
+happner-client: 12.2.0
+happner-cluster: 12.1.1
+
+