+++ packages/happner-cluster/RELEASES.md
@@ -350,6 +350,14 @@
 -----------------
   - TEN-95: $call discovers components and methods in cluster
 
+12.1.1 2022-04-12
+-----------------
+  - TEN-60: Wildcard-brokering
+  - TEN-42: Rest-discovery
+
+
+
+